requirements.txt
Django==5.0.4
django-environ==0.11.2
django4-background-tasks==1.2.9
djangorestframework==3.15.1
drf-yasg==1.21.7
environ==1.0
excel2json-3==0.1.6
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
python-dotenv==1.0.0

